Refactor user schemas import paths
- Updated import paths for user role schemas (Admin, Auditor, Teacher, Tutor, SuperAdmin) to point to the new location in "@/types/form/userSchemas".
- Removed the obsolete schemas.ts file as part of the restructuring.
